Fix issue #1 by using single remake file
Use single remake.yml by relying on the updated
parallelRemake package. Code is much cleaner and quicker,
and the interface is the same.

The example can now be generated and run from start to
finish with run_example_workflow(). write_example_workflow()
and setup_example_workflow() are also available. The main
README.md is updated to reflect this.